Escape double quotes in text values and use comma as delimiter
Escaping means doubling up the double quotes, i.e. "", so that each chunk of text is recognised properly using one double quote as the text delimiter. Using a comma as the field separator then means that the CSV can just be opened in Excel or imported into a Canvas course as is, rather than having a different separator and an extra step of a text import into Excel.